KRACOEUS-7051: move person extended attributes maintainable to common.impl
