+++ coeus-code/src/main/java/org/kuali/coeus/common/impl/person/attr/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Kuali Coeus Person Extended Attributes Components
+ * 
+ * 
+ * 
+ * @author Kuali Coeus Development Team
+ *
+ */
+package org.kuali.coeus.common.impl.person.attr;
