+++ coeus-code/src/main/java/org/kuali/kra/rules/KcPersonExtendedAttributesMaintenanceDocumentRule.java
@@ -2,7 +2,7 @@ package org.kuali.kra.rules;
 
 import org.apache.commons.lang3.StringUtils;
 import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributes;
-import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributesMaintainableImpl;
+import org.kuali.coeus.common.impl.person.attr.KcPersonExtendedAttributesMaintainableImpl;
 import org.kuali.coeus.sys.framework.rule.KcMaintenanceDocumentRuleBase;
 import org.kuali.kra.infrastructure.KeyConstants;
 import org.kuali.kra.rule.event.SaveCustomDataEvent;
