+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/docs/KcPersonExtendedAttributesMaintenanceDocument.xml
@@ -4,7 +4,7 @@
 
   <bean id="KcPersonExtendedAttributesMaintenanceDocument-parentBean" abstract="true" parent="KcMaintenanceDocumentEntry">
     <property name="businessObjectClass" value="org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributes"/>
-    <property name="maintainableClass" value="org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributesMaintainableImpl"/>
+    <property name="maintainableClass" value="org.kuali.coeus.common.impl.person.attr.KcPersonExtendedAttributesMaintainableImpl"/>
     <property name="maintainableSections">
       <list>
         <ref bean="KcPersonExtendedAttributesMaintenanceDocument-EditPerson"/>
