+++ coeus-it/src/test/java/org/kuali/kra/bo/KcPersonExtendedAttributesMaintainableImplTest.java
@@ -20,7 +20,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributes;
-import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributesMaintainableImpl;
+import org.kuali.coeus.common.impl.person.attr.KcPersonExtendedAttributesMaintainableImpl;
 import org.kuali.kra.maintenance.MaintenanceRuleTestBase;
 import org.kuali.rice.kns.document.MaintenanceDocument;
 import org.kuali.rice.kns.web.ui.Field;
