+++ coeus-it/src/test/java/org/kuali/kra/bo/KcPersonExtendedAttributesMaintenanceDocumentRuleTest.java
@@ -19,7 +19,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributes;
-import org.kuali.coeus.common.framework.person.attr.KcPersonExtendedAttributesMaintainableImpl;
+import org.kuali.coeus.common.impl.person.attr.KcPersonExtendedAttributesMaintainableImpl;
 import org.kuali.kra.maintenance.MaintenanceRuleTestBase;
 import org.kuali.kra.rules.KcPersonExtendedAttributesMaintenanceDocumentRule;
 import org.kuali.rice.core.api.util.RiceKeyConstants;
